half-working context switch?
I'm pretty sure the bug is in the _swi_handler, where it doesn't
properly save the user registers on the stack. Things are complicated by
the fact that the user-mode return address from the swi instruction is
placed in the Supervisor LR instead of the User LR. I feel like there
should be a really easy way to get the lr_svc value from system mode,
but idk.