read sensor data asynchronously
I was seeing weird behaviour on the track from doing the Getn after sending
off the query request. Sometimes the whole prompt would hang (presumably
from a missed byte) and sometimes we'd get extra bytes at the beginning
of the response. Instead, there's now a SensorReporterTask that
continuously tries to read from COM1 and updates the sensors whenever we
receive a value.